Crate rust_unix_socket_lengths

Navigating the path limits of UNIX domain sockets

This was sparked by Leonard Poettering’s tweet1 about a cool workaround for AF_UNIX socket path lengths. Though they vary from system to system, the largest in the wild seems to be 108 chars 2, with BSD-derivatives using 104 chars, and more obscure systems running even smaller3. This can turn out to be a real issue4 that can come up when working with XDG paths (e.g. $XDG_RUNTIME_PATH as mentioned by Leonard) or when dealing with hashes in paths.

They lay out the two ways to resolve issues around this:

Leverage procfs as an alternative location for a file descriptor

Use temp files and rename after the path has been used (checked is too gracious a term, given it’s simply a hard-limit)

Step-by-step

We start of here with a long path, which we can see in the output exceeds the maximum path length:

fn main() -> Result<()> {
    let sock_dir = "/tmp/application/lorem/ipsum/dolor/sit/amet/consectetur/adipiscing/elit/sed/do/eiusmod/tempor/incididunt";
    std::fs::create_dir_all(&sock_dir)?;

    let mut sock_path = PathBuf::from(&sock_dir);
    sock_path.push("server.sock");
    println!("path: {}\nlength: {}\nmaximum length:{}",
        sock_path.to_str().unwrap(),
        sock_path.as_os_str().len(),
        sockaddr_un_path_max());
    Ok(())
}

Run

Output:

path: /tmp/application/lorem/ipsum/dolor/sit/amet/consectetur/adipiscing/elit/sed/do/eiusmod/tempor/incididunt/server.sock
length: 116
maximum length:108

The methods to avoid running into the sockaddr_un path length limit differ between binding and connecting to a socket.

Binding a socket

To bind to a long path, we need to setup a symlinked directory to the parent directory of the socket at the end of the long path.

Setup the socket(2), then create the path that will serve as the address for the socket to be bound, to within the short directory.

bind(2) will create the socket file at the path specified in our sockaddr_un2’s sun_path. Once we have this, we can rename the file to our obscenely long directory:

/tmp/short-directory/application.sock -> /var/run/obscenely/.../long/application.sock

fn main() -> Result<()> {
    let sock_dir = LONG_PATH;
    std::fs::create_dir_all(&sock_dir)?;

    let mut sock_path = PathBuf::from(&sock_dir);
    sock_path.push("server.sock");

    let random_dir = "/tmp/tempdir-1652813977685043874";
    symlink(&sock_dir, &random_dir)?;

    let sock_fd = socket(libc::AF_UNIX, libc::SOCK_STREAM, 0);

    let sock_temp_path = format!("{}/{}", random_dir, "server.sock");
    let (address, address_len) = sockaddr_un_from_path(&sock_temp_path);

    bind_and_listen(sock_fd, address, address_len);
    rename(&sock_temp_path, sock_path)?;
    Ok(())
}

Run

Connecting a socket

To connect to a long path, we setup our socket(2) and then open the long path to the socket server socket using the O_PATH option. It’s worth noting that Rust does not allow for opening files with an all-false or empty access mode

use std::fs::file;

// This will error with io::ErrorKind::InvalidInput
let long_socket = File::options()
    .custom_flags(libc::O_PATH)
    .open(&long_socket_path)?;

// This works
let long_socket = File::options()
    .read(true)
    .custom_flags(libc::O_PATH)
    .open(&long_socket_path)?;
Run

This is only unintuitive because O_PATH is the only exception to the rule that a file most be opened with some access mode. You can call libc::open(cstr, libc::O_PATH) directory on UNIX systems without issues. Under normal circumstances, this rule is sensible and hides a potential foot-gun.

We can use the fd from the open(..., O_RDONLY | O_PATH) to build the path to the procfs fd:

/proc/self/fd/<fd>

We assume that this path will be shorter than 108/SUN_PATH chars, which allows us to use it as the sockaddr_un.’s sun_path field. With that, we’re able to construct a valid sockaddr_un and connect(2) without any trouble.

Demo:

fn main() -> Result<()> {
    let sock_dir = LONG_PATH;
    std::fs::create_dir_all(&sock_dir)?;

    let mut sock_path = PathBuf::from(&sock_dir);
    sock_path.push("server.sock");

    let sock_fd = socket(libc::AF_UNIX, libc::SOCK_STREAM, 0);
    // This will fail unless the socket has been bound already. Run `cargo test` with the
    // checked out version of this project to see that it works.
    let dest_sock_fd = File::options()
        .read(true)
        .custom_flags(libc::O_PATH)
        .open(&sock_path)?;
    let dest_sock_fd = dest_sock_fd.as_raw_fd();

    let procfs_sock_path = PathBuf::from(format!("/proc/self/fd/{dest_sock_fd}"));
    let (address, address_len) = sockaddr_un_from_path(&procfs_sock_path);

    connect(sock_fd, address, address_len);
    Ok(())
}

Run

https://twitter.com/pid_eins/status/1524666972880916480 ↩

Struct definition in the Linux kernel: https://elixir.bootlin.com/linux/v5.17.8/source/include/uapi/linux/un.h#L9 ↩

StackOverflow post discussing these sizes with great references: https://unix.stackexchange.com/questions/367008/why-is-socket-path-length-limited-to-a-hundred-chars ↩

Podman ran into this and resolved the connect(2) issue: https://github.com/containers/podman/issues/8798 ↩

Functions

bind_to_long_path

Bind an AF_UNIX stream socket to path which exceeds system sockaddr_un path limit

connect_to_long_path

generate_base_name

Generates a base-name in the format of sockets-<UNIX nanosecond timestamp>

main

prepare_too_long_path

Creates a path that is longer than the system’s configured sockaddr_un->sun_path length, using “lorem ipsum”.

sockaddr_un_from_path⚠

sockaddr_un_path_max

Programmatically get the size of the sun_path sockaddr_un field which varies from 92 to 108 bytes.

sockaddr_un_path_offset